+++ apis/python/src/tiledbsoma/_dask/util.py
@@ -21,8 +21,7 @@ from numpy.typing import NDArray
 from somacore.options import SparseNDCoord, SparseNDCoords
 from typing_extensions import TypeAlias
 
-from tiledbsoma.options import SOMATileDBContext
-from tiledbsoma.options._soma_tiledb_context import ConfigVal
+from tiledbsoma._soma_context import SOMAContext
 
 ChunkSize = Union[int, tuple[Union[int, None], int]]
 JoinIDs: TypeAlias = NDArray[int64]
@@ -38,7 +37,7 @@ class SOMADaskConfig(TypedDict, total=False):
     """
 
     chunk_size: ChunkSize
-    tiledb_config: dict[str, ConfigVal]
+    tiledb_config: dict[str, str | float]
 
 
 def chunk_ids_sizes(joinids: JoinIDs, chunk_size: int, dim_size: int) -> tuple[list[JoinIDs], list[int]]:  # noqa: ARG001
@@ -57,7 +56,7 @@ def chunk_ids_sizes(joinids: JoinIDs, chunk_size: int, dim_size: int) -> tuple[l
     return chunk_joinids, chunk_sizes
 
 
-def make_context(tiledb_config: dict[str, Any]) -> SOMATileDBContext:
+def make_context(tiledb_config: dict[str, Any]) -> SOMAContext:
     """Create and cache ``SOMATileDBContext``s within Dask worker processes.
 
     This wrapper just flattens the ``tiledb_config`` ``dict`` into hashable ``tuple``s (for use with ``cache``).
@@ -66,12 +65,12 @@ def make_context(tiledb_config: dict[str, Any]) -> SOMATileDBContext:
 
 
 @cache
-def _make_context(tiledb_configs: tuple[tuple[str, Any], ...]) -> SOMATileDBContext:
+def _make_context(tiledb_configs: tuple[tuple[str, Any], ...]) -> SOMAContext:
     """Create and cache ``SOMATileDBContext``s within Dask worker processes.
 
     ``tiledb_config`` is conceptually a ``dict``, but flattened into hashable ``tuple``s here, for use with ``cache``.
     """
-    return SOMATileDBContext(tiledb_config=dict(tiledb_configs))
+    return SOMAContext.create(config=dict(tiledb_configs))
 
 
 def coord_to_joinids(coord: SparseNDCoord, n: int) -> JoinIDs: