+++ apis/python/src/tiledbsoma/options/_soma_tiledb_context.py
@@ -5,9 +5,9 @@
 from __future__ import annotations
 
 import datetime
-import functools
 import threading
 import time
+import warnings
 from collections.abc import Mapping
 from concurrent.futures import ThreadPoolExecutor
 from typing import Any, Literal, Optional, Union, cast
@@ -16,6 +16,7 @@ from somacore import ContextBase
 from typing_extensions import Self
 
 from tiledbsoma import pytiledbsoma as clib
+from tiledbsoma._soma_context import SOMAContext
 from tiledbsoma._types import DataProtocol, OpenTimestamp
 from tiledbsoma._util import ms_to_datetime, to_timestamp_ms
 
@@ -46,6 +47,24 @@ ReplaceConfig = dict[str, Optional[ConfigVal]]
 """Replacing a value with ``None`` serves to delete that key."""
 
 
+def _update_context_and_timestamp(
+    context: SOMAContext | SOMATileDBContext | None, tiledb_timestamp: OpenTimestamp | None
+) -> tuple[SOMAContext, OpenTimestamp | None]:
+    """Helper function that returns the context and timestamp.
+
+    This function can be removed after the :class:`SOMATileDBContext` is deprecated.
+    """
+    if isinstance(context, SOMATileDBContext):
+        if tiledb_timestamp is None and context.timestamp_ms is not None:
+            tiledb_timestamp = context.timestamp_ms
+        return context._to_soma_context(), tiledb_timestamp
+    if context is None:
+        if not SOMAContext.has_default():
+            SOMAContext.set_default()
+        return SOMAContext.get_default(), tiledb_timestamp
+    return context, tiledb_timestamp
+
+
 def _default_config(override: ConfigMap) -> ConfigDict:
     """Returns a fresh dictionary with TileDB config values.
 
@@ -55,12 +74,6 @@ def _default_config(override: ConfigMap) -> ConfigDict:
     cfg: ConfigDict = {"sm.mem.reader.sparse_global_order.ratio_array_data": 0.3}
     cfg.update(override)
     return cfg
-
-
-@functools.cache
-def _default_global_native_context() -> clib.SOMAContext:
-    """Lazily builds a default SOMAContext with the default config."""
-    return clib.SOMAContext({k: str(v) for k, v in _default_config({}).items()})
 
 
 def _maybe_timestamp_ms(input: OpenTimestamp | None) -> int | None:
@@ -83,7 +96,7 @@ class SOMATileDBContext(ContextBase):
     a new ``SOMATileDBContext`` with new values.
 
     Lifecycle:
-        Maturing.
+        Deprecated.
     """
 
     def __init__(
@@ -149,6 +162,8 @@ class SOMATileDBContext(ContextBase):
                 provided, a new ThreadPoolExecutor will be created with
                 default settings.
         """
+        warnings.warn("SOMATileDBContext is deprecated. Use SOMAContext instead.", DeprecationWarning, stacklevel=1)
+
         if tiledb_ctx is not None and tiledb_config is not None:
             raise ValueError("only one of tiledb_config or tiledb_ctx may be set when constructing a SOMATileDBContext")
 
@@ -177,6 +192,15 @@ class SOMATileDBContext(ContextBase):
 
         self._native_context: clib.SOMAContext | None = None
         """Lazily construct clib.SOMAContext."""
+
+    def _to_soma_context(self) -> SOMAContext:
+        if self._native_context is not None:
+            return SOMAContext(
+                self._native_context,
+                threadpool=self.threadpool,
+                _dont_call_this_use_create_instead="tiledbsoma-internal-code",
+            )
+        return SOMAContext.create(self.tiledb_config, self.threadpool)
 
     @property
     def timestamp_ms(self) -> int | None:
@@ -210,8 +234,12 @@ class SOMATileDBContext(ContextBase):
                     cfg = self._tiledb_ctx.config().dict()
                     self._native_context = clib.SOMAContext({k: str(v) for k, v in cfg.items()})
                 else:
-                    # The user did not provide settings so create a default
-                    self._native_context = _default_global_native_context()
+                    # The user did not provide settings so create a default - use SOMAContext to ensure there is
+                    # a single internal default.
+                    if not SOMAContext.has_default():
+                        SOMAContext.set_default()
+                    soma_context = SOMAContext.get_default()
+                    self._native_context = soma_context._handle
 
         # else SOMAContext already exists
         return self._native_context
@@ -297,7 +325,7 @@ class SOMATileDBContext(ContextBase):
                 A threadpool to replace the current threadpool with.
 
         Lifecycle:
-            Maturing.
+            Deprecated.
 
         Examples:
             >>> context.replace(timestamp=1_512_658_800_000)  # UNIX millis
@@ -387,24 +415,3 @@ class SOMATileDBContext(ContextBase):
     def is_tiledbv3_uri(self, uri: str) -> bool:
         """Return True if the URI will use `tiledbv3` semantics."""
         return self.data_protocol(uri) == "tiledbv3"
-
-
-def _validate_soma_tiledb_context(context: Any) -> SOMATileDBContext:  # noqa: ANN401
-    """Returns the argument, as long as it's a ``SOMATileDBContext``, or a new
-    one if the argument is ``None``. While we already have static type-checking,
-    a few things are extra-important to have runtime validation on.  Since it's
-    easy for users to pass a ``tiledb.Ctx`` when a ``SOMATileDBContext`` is
-    expected, we should offer a helpful redirect when they do.
-    """
-    if context is None:
-        return SOMATileDBContext()
-
-    if TILEDB_EXISTS and isinstance(context, TileDBCtx):
-        raise TypeError(
-            "context is a tiledb.Ctx, not a SOMATileDBContext -- please wrap it in tiledbsoma.SOMATileDBContext(...)",
-        )
-
-    if not isinstance(context, SOMATileDBContext):
-        raise TypeError(f"context is not a SOMATileDBContext: got {type(context)}")
-
-    return context